Added parsing for RACE command.  In the process, refactored parsing integer values in ranges, which will come in handy in lots of places.
Finally, adjusted the parseFailure() method in BaseKnownCommandParser to
be generic, so it could be used as a return value from methods
with different types.  parseFailure() really has the goal of throwing an
exception, but the compiler doesn't pick up on this (requires global
analysis that it doesn't do), meaning that it reports that certain code
paths don't return a value.